+++ src/Outrun.hs
@@ -1,5 +1,8 @@
 module Outrun (module Outrun) where
 
+-- | This module contains all essential functions
+-- | for the Outrun game
+
 import           Data.List
 import           Data.List.HT
 import           Graphics.Gloss
@@ -57,7 +60,7 @@ projectRoadLines = projectRoadLines' (0, 0, 0, 0)
 
         0 -> [] -- Zero render distance
 
-        _ -> projectRoadLines' (x', dx', y', dy') cam' rest ++ [lineProj']
+        _ -> projectRoadLines' (x', dx', y', dy') cam' rest ++ [lineProj]
           where
             -- Applying change in curve and pitch values
             x' = x + dx
@@ -67,11 +70,8 @@ projectRoadLines = projectRoadLines' (0, 0, 0, 0)
             dy' = dy + roadLinePitchRate line
 
             -- Projecting road line
-            lineProj = projectRoadLine cam (shiftRoadLine (0, y, 0) line)
-            lineProj' = lineProj {
-              projectedPosition =
-                shiftPoint (x, 0) (projectedPosition lineProj)
-              }
+            lineProj = shiftProjected (x, 0)
+              (projectRoadLine cam (shiftRoadLine (0, y, 0) line))
 
             -- Reducing number of road lines to render
             rendDist = cameraRenderDistance cam
@@ -155,7 +155,7 @@ drawTerrainSegment
   -> Color -- Color of the terrain segment
   -> Picture
 drawTerrainSegment (x, y) p2 col =
-   -- Drawing the sides of the terrain
+  -- Drawing the sides of the terrain
   drawRoadSegment (x, -1000) 10000 p2 10000 (dark col)
 
   -- Drawing the terrain road segment
@@ -214,10 +214,7 @@ drawRacingTrack cam segmentDrawers track roadObjs =
 
         -- Drawing road segment between two road lines (could include terrain segement)
         roadSegment =
-          foldr ((<>) . ($color) . ($farW) . ($farP) . ($nearW) . ($nearP) . ($index))
-            -- do not know how to make this look better for now
-            blank
-            segmentDrawers
+          pictures (map (\f -> f index nearP nearW farP farW color) segmentDrawers)
 
         -- Drawing road objects between two road lines
         drawnRoadObjects = pictures $
@@ -344,19 +341,19 @@ updateGame dt state = updatedGameState
     (cx, cy, cz) = cameraPosition cam
     (px, __, pz) = roadObjectPosition player
 
-    -- Shows whether the player is on the track or not
-    isOnTrack = (defaultRoadSegmentWidth - abs px) >= 0
-
     -- Getting road line nearest to the player from behind
     (playerRL : _) = dropWhile ((< pz) . getZ . roadLinePosition) track
 
     curveRate = roadLineCurveRate playerRL
     pitchRate = roadLinePitchRate playerRL
 
+    -- Shows whether the player is on the track or not
+    isOnTrack = (roadLineWidth playerRL - abs px) >= 0
+
     -- Get x- and z-direcred movement scalar
-    acc         = 0.0025
+    acc         = if isOnTrack then 0.0025 else 0.01
     centrifugal = 0.3
-    maxZSpeed   = if isOnTrack then 320 else 100
+    maxZSpeed   = if isOnTrack then 320 else 50
     maxXSpeed   = 200
 
     ddx = maxXSpeed * fromIntegral (iRight - iLeft)
@@ -366,8 +363,15 @@ updateGame dt state = updatedGameState
     spdZ' = max 0 (approachSmooth spdZ ddz acc - pitchRate / 100)
     --      ^^^ prohibits player from going backwards
 
+    -- Retuns value
+    dxClamp = let
+        off = defaultRoadSegmentWidth * 2
+        bound = roadLineWidth playerRL + off
+      in clamp (-bound) bound px - px
+
     -- Update player position
-    updatedPlayer = shiftRoadObject (spdX', 0, spdZ') player
+    updatedPlayer =
+      shiftRoadObject (spdX' + dxClamp, 0, spdZ') player
 
     -- Update camera position
     cameraSmoothness = 0.5
@@ -409,7 +413,7 @@ outrunDisplayTest :: RacingTrack -> IO ()
 outrunDisplayTest track =
   display window white (drawRacingTrack cam [] track [])
   where
-    cam    = Camera (0, 1500, 0) (1024, 768) 0.3 100 0.75
+    cam    = Camera (0, 1500, 0) (1024, 768) 0.3 100 0.75 0
     window = InWindow "Debug -- sample track" (1024, 768) (0, 0)
 
 
@@ -417,7 +421,7 @@ outrunPlayTest :: RacingTrack -> IO ()
 outrunPlayTest track =
   play window white 60 initState drawGame handleInput updateGame
   where
-    cam    = Camera (0, 1000, 0) (1024, 768) 0.4 200 0.75
+    cam    = Camera (0, 1000, 0) (1024, 768) 0.4 100 0.75 0
     window = InWindow "Debug -- sample track ride" (1024, 768) (0, 0)
 
     initState = GameState [] cam (infRacingTrack track) player
@@ -446,6 +450,6 @@ outrunPlayTest track =
 
     drawGame :: OutrunGameState -> Picture
     drawGame (GameState _ cam track player) =
-      drawRacingTrack cam [terrainPic, roadSegment] track [player]
+      drawRacingTrack cam [roadSegment] track [player]
       <>
       text (show (roadObjectVelocity player))