+++ ChangeLog.md
@@ -1,9 +1,10 @@
-# Changelog for HaskellGame
+# Changelog for haskell-outrun
 
 0.1.1.0 - Added working road (lines) retro scrolling
 0.1.3.0 - Added infinite racing track
 0.1.4.0 - Added curved tracks and track building functionality
 0.1.5.0 - Added hills and interpolation functions
 0.1.5.1 - Added pitched tracks
+
 0.2.0.1 - Added projecting functionality for road objects
           Refactored project architecture